Refactor: global.css 파일 위치 수정 (src/app -> src/styles)
