+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import './globals.css';
+import '../styles/globals.css';
 
 import { Inter } from 'next/font/google';
 
@@ -15,7 +15,6 @@ const RootLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <html lang="en">
       <body className={inter.className}>
-        {/* {children} */}
         <Provider>{children}</Provider>
       </body>
     </html>
